Move the base item-sheet file
